Bug fix - Logger cannot be created in +initialize method
Summary:
We need to be really careful about how we work with dependencies that have `+initialize` methods.

In this case we were logging when `GateKeeperManager` wasn't configured, but without configuring we cannot satisfy a `Logging` dependency. Similarly, if we create the concrete logger in the `+initialize` method then we don't have the opportunity to its dependency on `Settings`.

Additionally this diff moves the configuration of `Settings` in `ApplicationDelegate` closer to its declaration.

Reviewed By: Oliverccccct

Differential Revision: D28258855

fbshipit-source-id: 2aa251899272e294f3b84dc8651bdd721c0a2124