Exposing BridgeAPI, removing some unused types from internal umbrella header
Summary:
Transitioning to Swift demands that we have clear boundaries between modules.

Use of types in internal umbrella-header `FBSDKCoreKit+Internal.h` should not occur from other modules but it does.

This exposes a few of those types with warnings not to use them so that we can begin decoupling (or in this case, making the coupling explicit) between modules.

Reviewed By: jamestouri

Differential Revision: D28145292

fbshipit-source-id: af81a1cceacade8d0595506a5d217e3ed60b2799