Add Basics Podspec
Summary: The issue is that we have a mismatch between how basics is consumed. In terms of BUCK, SPM, and Xcodebuild it's a true module. In the case of CocoaPods it's a "subspec" which is not actually a module. This means we have differences in import statements, and general inconsistencies between our build processes which slows development and is a general pain to manage.

Reviewed By: jawwad

Differential Revision: D28522118

fbshipit-source-id: 953629bd5f365af1fcc064e0f7b6fcf0a07ab48d