Remove OCMock from FBSDKProfileTests.m
Summary:
The existing code that was removed was doing 2 things. 1. It was verifying the `request.graphPath` and 2. it was verifying that `loadProfileWithToken:completion:graphRequest:` would be called when `loadProfileWithToken:completion:` was called.

For #1. I moved the code that creates the graph path into a new `graphPathForToken:` method which the existing `loadProfileWithToken:completion:` will call. The test will call the `graphPathForToken:` method directly to get this information (instead of reaching into the `loadProfileWithToken:completion:graphRequest:` method as it was doing before.
For #2. I deleted the code since it really isn't necessary to verify that `loadProfileWithToken:completion:graphRequest:` is called when `loadProfileWithToken:completion:` is called since that's all the method does (after creating the graph path and graph request, so it will always be called). I think it was being verified before just because OCMock needed to call the method anyway in order to grab the request object to check the path.

Also the `testGraphPathForProfileLoadWithFriendsPermission` test was actually verifying the graphPath twice. Once in the method itself and once with the `verifyGraphPath` call which did the exact same thing. Removed the extra check in the method itself.

Reviewed By: joesus

Differential Revision: D28616761

fbshipit-source-id: 41a8a3a0e9ad719382be8de95aafe847eaffdfd3