Backfill Unit Tests - FBSDKEventBinding
Summary: This was a little more involved than planned. Needed to stop stubbing AppEvents with OCMock so instead am providing it as a dependency via its owner, AppEvents. It doesn't make sense to create a new concrete dependency for AppEvents so instead introducing that ApplicationDelegate (current clearing house for all dependencies) and injecting it down the chain.

Reviewed By: jawwad

Differential Revision: D28655902

fbshipit-source-id: 3835c88427fac171d0836fae819634c27ab92970