Refactor FBSDKProfileTests to be a subclass of XCTestCase instead of FBSDKTestCase
Summary:
This replaces calls to the stubGraphRequestWithResult: method. Updated to manually call the capturedCompletionHandler on a testGraphRequest block. There were 2 cases in which the normal completionHandler wouldn't be called. In those cases added an XCTestExpectation to ensure that the XCTAsserts in the completion block of the loadProfileWithToken: method were getting called.

Also set the indentation to be 2 spaces at the project level in the TestTools project.

Reviewed By: joesus

Differential Revision: D28651658

fbshipit-source-id: f8ce907aa7732143ef053c0f87f8b23af7b045a4