Refactor TimeSpentData usage
Summary:
When `ApplicationDelegate` configures `AppEvents` it will pass a factory for creating objects the help `AppEvents` will use for for timing sessions (they conform to `SourceApplicationTracking` and `TimeSpentRecording`).

This object, implemented by the concrete type `TimeSpentData`, needs to be informed when the source application changes but has not other requirements outside of `AppEvents`.

Therefore we can store it on `AppEvents` and use `AppEvents` as a façade for `TimeSpentData` by conforming to `SourceApplicationTracking`. This means no other types need to know about `TimeSpentData` or `TimeSpentRecording`.

Reviewed By: jawwad

Differential Revision: D28693011

fbshipit-source-id: 0ae7faa43b2bac9fcddf25f93627cfaa5e64d833