Removing OCMock from FBSDKFriendFinderDialogTests
Summary:
The main motivation for this is that mocking `BridgeAPI` from these tests was presenting a blocker to refactoring the BridgeAPITests themselves.

More generally, this cleanup is necessary because we are working towards a rewrite in Swift and OCMock will require us to expose all of our rewritten types to the ObjC runtime which is a non-starter.

This required abstracting creating of a `GamingServiceController` from the `FriendFinderDialog` by introducing a new factory type `GamingServiceControllerCreating` and concrete implementation `GamingServiceControllerFactory` which returns the new concrete type `GamingServiceController` which conforms to a new abstraction `GamingServiceControllerProtocol`.

Reviewed By: Alaeldin513

Differential Revision: D29032601

fbshipit-source-id: 4743cc2c3b7969871acca31d3f622347c30622b9