Removing OCMock from GamingImageUploaderTests 1/n
Summary:
The main motivation for this is that mocking `BridgeAPI` from these tests was presenting a blocker to refactoring the BridgeAPITests themselves.

More generally, this cleanup is necessary because we are working towards a rewrite in Swift and OCMock will require us to expose all of our rewritten types to the ObjC runtime which is a non-starter.

The abstractions used to do this are introduced in D29032601

Reviewed By: jawwad

Differential Revision: D29038155

fbshipit-source-id: 4cdc80f4bade5f946c7b7b086dd615f233d6ccb7