Removing OCMock from AppEventsTests 2/n
Summary:
More of the same. This required me to provide a backdoor to set the singleton instance of AppEvents in tests.

The alternative to this was to create an underlying instance method for each class method and test those. This seemed better since it allows us to continue testing the public interface which remains class-based for now.

Reviewed By: linmx0130

Differential Revision: D29114153

fbshipit-source-id: df54c395e06bd9659f81d3a3bba6e2aa0492b846