{FriendFinder]Remove Token Check
Summary: We got reports that the developers trigger to show the friend finder dialog is unresponsive on iOS 14 when tracking is enabled. This check was the cause of that issue so we're removing the check since the dialog is an app switch and the user would need to be logged in.

Differential Revision: D29076864

fbshipit-source-id: d31e7cfd7d1648c0a108351d4da92b99f6af89ad