Attempt to generate xccode project for FBSDKCoreKit_Basics
Summary:
Xcode project files are difficult to read, work with, and collaborate on. This is why companies typically generate their project files using tools like BUCK, CocoaPods, Swift Package Manager, and Xcodegen instead of trying to manage them in source control.

Xcodegen is a tool that allows us to generate our project files on demand from a configuration file. The plan is to move entirely to Xcodegen and to remove the checked-in project files from source control.

Differential Revision: D29091884

fbshipit-source-id: ea66660f690b6b970dc258e9f107f9d03cff8b7d