Xcodegen - Add FBSDKCoreKitTests and TestTools
Summary:
Xcode project files are difficult to read, work with, and collaborate on. This is why companies typically generate their project files using tools like BUCK, CocoaPods, Swift Package Manager, and Xcodegen instead of trying to manage them in source control.

Xcodegen is a tool that allows us to generate our project files on demand from a configuration file. The plan is to move entirely to Xcodegen and to remove the checked-in project files from source control.

The plan to tackle FBSDKCoreKit.xcodeproj is to add all the target definitions independently and then replace the project file.

This diff:
Adds a definition for the FBSDKCoreKitTests target and a new project.yml for TestTools.xcodeproj

fbshipit-source-id: e1ab6293f484bf3f0ed407fa2f7b56b69d4c69ae