Adding FBSDKCoreKit_TV(-Dynamic) Target
Summary:
Xcode project files are difficult to read, work with, and collaborate on. This is why companies typically generate their project files using tools like BUCK, CocoaPods, Swift Package Manager, and Xcodegen instead of trying to manage them in source control.

Xcodegen is a tool that allows us to generate our project files on demand from a configuration file. The plan is to move entirely to Xcodegen and to remove the checked-in project files from source control.

The plan to tackle FBSDKCoreKit.xcodeproj is to add all the target definitions independently and then replace the project file.

This diff:
Adds definitions for the FBSDKCoreKit_TV and FBSDKCoreKit_TV-Dynamic targets

Differential Revision: D29214198

fbshipit-source-id: 1c496404af88ad46d932159bdf8ce7fb5a2deb63