Xcodegen - Replacing existing CoreKit project and schemes with generated
Summary:
Ran:
```
cd FBSDKCoreKit && xcodegen generate
cd TestTools && xcodegen generate
```
The project files were compared in previous diffs so any discrepancies are expected. See test plan from previous diffs in the stack for explanation of specific changes.

Once we are sure these new project files work for us we can configure everything so that they are generated in CI and then delete them.

Note: Got rid of standalone FBSDKCoreKitTests scheme because it makes more sense to me to run it from FBSDKCoreKit-Dynamic. Open to discussion on this. It's easy to add.

Differential Revision: D29232163

fbshipit-source-id: c8b8d6e47c2d50607979490ccec8d4821549cdde