Xcodegen - adding aggregate targets
Summary:
Xcode project files are difficult to read, work with, and collaborate on. This is why companies typically generate their project files using tools like BUCK, CocoaPods, Swift Package Manager, and Xcodegen instead of trying to manage them in source control.

Xcodegen is a tool that allows us to generate our project files on demand from a configuration file. The plan is to move entirely to Xcodegen and to remove the checked-in project files from source control.

The plan to tackle FBSDKCoreKit.xcodeproj is to add all the target definitions independently and then replace the project file.

This diff adds definitions for:
FBSDKCoreKit-Universal
FBSDKCoreKit_TV-Universal

Reviewed By: jamestouri

Differential Revision: D29232292

fbshipit-source-id: bddb6f4e75f3b0c27465120e60de0019c194b123