Xcodegen FBSDKShareKit
Summary:
Xcode project files are difficult to read, work with, and collaborate on. This is why companies typically generate their project files using tools like BUCK, CocoaPods, Swift Package Manager, and Xcodegen instead of trying to manage them in source control.

Xcodegen is a tool that allows us to generate our project files on demand from a configuration file. The plan is to move entirely to Xcodegen and to remove the checked-in project files from source control.

This diff adds a project configuration file for FBSDKShareKit.

Reviewed By: jamestouri

Differential Revision: D29335943

fbshipit-source-id: ecde5dc45a9fdbc96ae5d945719d51e0418b1137