Abstract dependency of SKAdNetworkReporter from AppEvents
Summary:
A new `AppEventsReporter` protocol is introduced to abstract the behavior of `SDKAdNetworkReporter` and `AEMReporter`. In this diff, `AppEvents` will only access the methods on the protocol rather than the full `SDKAdNetworkReporter`.

Also there's a bug: `SKAdNetworkReporter` will only be configured on iOS 14+ before. However this feature is introduced in iOS 11.3. I added the check back. The tests around `SKAdNetworkReporter` in `AppEvents` are backfilled thanks to this dependency abstraction.

Reviewed By: joesus

Differential Revision: D29643096

fbshipit-source-id: 9b19b8dbb9db7cf981acfbece5bd0673772d53d4