Script cleanup
Summary:
Cleans up xcodegen script to check for expected version.

This should prevent unexpected changes from people with different versions. At some point we probably want to use some other means of versioning this. Maybe a Swift command line tool would be easiest. In the short term this solves the issue, we can revisit later if it becomes a pain to manage versions this way.

Reviewed By: jawwad

Differential Revision: D29672533

fbshipit-source-id: 3feba368f0cb32cf9909ba9aeb1d779f967413da