Add minSize and maxSize Parameters
Summary:
Adding the minSize and maxSize parameters to the context choose dialog content. This will allow for developers to specify the number of users in a given context. If a context size falls outside of the range then it will not be shown in the context choose dialog.

Note:
Its not really important right since we aren't using it on the server side. We are only working with threads context  source types for now.

Reviewed By: joesus

Differential Revision: D29615421

fbshipit-source-id: 3fbe42b136757d58fa9683886dc8794c7bf63e92