Fixed issue with 3 tests failing when running tests from the central time zone
Summary:
3 tests in PaymentProductRequestorTests were failing when running from the central time zone

The difference was the following:
transactionDate: Optional("0001-12-31 18:09:24-055036") vs
transactionDate: Optional("0001-12-31 16:07:02-075258")

The tests were hard coded to check "0001-12-31 16:07:02-075258" which is what the date formatter returned in PST, however in CST it would return 0001-12-31 16:07:02-075258.

Updated tests to generate the expected string using a date formatter that should generate the correct expected string regardless of time zone.

https://pxl.cl/1LFL3

Reviewed By: joesus

Differential Revision: D29704272

fbshipit-source-id: 83bb27f9323851c4ed01aafcd657a26bb86a35d4