Eliminate the FBSDKCopying protocol
Summary:
The trigger for removing this is that it is imported in other Kits (especially in Gaming Services via the FBSDKCoreKitInternalImport.h which we want to eliminate).

However it looks like this protocol is not really needed either.

The FBSDKCopying protocol looks like it was added as a convenience for declaring conformance to NSCopying and NSObject in a single protocol.

Other than this I fail to understand the purpose of it. It declares a copy method but the project continues to compile if that declaration is removed.

This was added in 2014 with D1748854. The commit message for that commit seems to indicate that this protocol does a bit more than just declaring conformance to NSCopying, NSObject but I still don't follow what the original inconvenience was that was mentioned in the diff. I believe it is safe to remove.

Reviewed By: joesus

Differential Revision: D29739226

fbshipit-source-id: 30190f5365c6972aeb7f4e9e14859686753e152a