Disable Analyze during build to speed up local builds
Summary:
To disable analyze during build the build setting RUN_CLANG_STATIC_ANALYZER needs to be set to NO. This shows up as "Analyze During 'Build'" in Xcode build settings.

Analyze during build isn't disabled via the scheme as I originally thought. Disabling it in the scheme disables the ability to analyze completely.

The other benefit is that any compile errors are shown clearly in the sidebar instead of just showing "Command AnalyzeShallow failed with a nonzero exit code". Here is an actual error:

{F632681421}

Here is what it would show before:

{F632681498}

Reviewed By: joesus

Differential Revision: D29748087

fbshipit-source-id: 5813ed2f3b690f28964feea0c451fe274d85a14e