Remove FBSDKCoreKitInternalImport.h from GamingImageUploader
Summary:
Makes `FBSDKGraphRequestConnectionFactory` public. Thought about just making a new wrapper in GamingServices but it seemed fine to expose it as a convenience from CoreKit since most of the time the default will be to use the concrete `FBSDKGraphRequestConnection` type.

It will still be easy to switch over to a custom factory if gaming ever needs to provide a custom type that conforms to `GraphRequestProtocol`.

Reviewed By: jawwad

Differential Revision: D29847768

fbshipit-source-id: 4ddf8b5378bf2ee6734e1f52628fbadaaab8ea64