Convert FBSDKAppLinkResolverRequestBuilderTests.m to Swift
Summary:
These tests were actually passing incorrectly in their current state.

request.parameters was empty and so the .location callon the result of the rangeOfString call was returning 0 for both tests.

NSNotFound is defined as NSIntegerMax so both tests were passing because 0 != NSIntegerMax.

Also in both tests it was checking for "iphone" which looks like a copy/paste error since one test was for iPhone and the other for iPad.

The request.graphPath was the thing that actually contained the iphone or ipad string.

For example in the ipad case request.graphPath was: ?fields=app_links.fields(ios,ipad)&ids=
And in the iphone case it was: ?fields=app_links.fields(ios,iphone)&ids=

Reviewed By: joesus

Differential Revision: D29857285

fbshipit-source-id: 582ca79a550bd2caefd0c55999d44be3079c90fb