Fix indentation issues across the iOS SDK using swiftformat
Summary:
SwiftFormat is a tool with which Swift code can be automatically formatted. Added a basic .swiftformat config to fix indentation issues across the iOS SDK using it.

Ran it with `swiftformat .`

Note: This is not Apple's swift-format tool (https://github.com/apple/swift-format) it is a tool that existed much earlier than that (https://github.com/nicklockwood/SwiftFormat).

Note: There is an issue in which it will de-indent protocol declarations when they are listed on a new line. There were 3 such files that had this issue. Surrounded that part with `// swiftformat:disable indent` and `// swiftformat:enable indent` so that formatting wouldn't be disabled on the file completely.

Reviewed By: joesus

Differential Revision: D29977861

fbshipit-source-id: d2570ce96d347cb4b3c7d295cb2a50d8cf84c3f3