Fix carthage archive step failing due to it creating the file with name FBAEMKit instead of FBSDKCoreKit
Summary:
It looks like Carthage takes the first framework it finds and uses it as the name of the zip file to create.

This was previous being created as FBSDKCoreKit.framework.zip and so the mv command was failing since it was now being created as FBAEMKit.framework.zip.

I found that the full path can be passed into the output option to give it the name to create the file as. This also eliminates the need for renaming.

Note: If a name is passed in to "carthage archive" for example "carthage archive FBSDKCoreKit" then it only archives that framework.

carthage archive --help doesn't show options so here they are for reference:
https://github.com/Carthage/Carthage/blob/master/Source/carthage/Archive.swift#L17-L35

Reviewed By: joesus

Differential Revision: D30052933

fbshipit-source-id: 4572d0304eea416d598091c3a78cbcded552295d