Refactor FBSDKGamingContext
Summary:
Adding unit tests and changing the semantics a bit.

Main change is getting rid of the side effect where setting a nil identifier nils out a shared instance. The code now manages whether or not a shared instance is available since it's a valid state to not have a current context.

Reviewed By: jamestouri

Differential Revision: D30057345

fbshipit-source-id: 45d05826cd51547e8fb46f17b6287439b7d9d2e1