Refactor how CocoaPods are released to account for the increasing number of dependencies
Summary:
Previously CocoaPods were pushed in 4 tiers of dependencies.
Base: FBSDKCoreKit
Dependent: FBSDKLoginKit FBSDKShareKit
Top Level: FBSDKGamingServicesKit FBSDKTVOSKit
Aggregate: FacebookSDK

However due to the addition of FBSDKCoreKit_Basics, FBAEMKit, LegacyGamingServices and FacebookGamingServices there are now 7 levels of dependencies

I've removed the individual steps and have placed all of the commands in a single scripts/publish_cocoapods.sh file

Ideally if a single command fails due to a caching issue we could retry just that command. I'll have to figure out how to do that with bash or will need to update this to be a python script for that. In any case, rerunning the complete script is ok since pushed Pods will fail with
`[!] Unable to accept duplicate entry for: FBSDKCoreKit_Basics (11.1.0)"`

Reviewed By: joesus

Differential Revision: D30078664

fbshipit-source-id: b1cde7bcfe309f324f8c1626193da719abcbf104