Enable the indentation_width opt_in rule in SwiftLint
Summary:
The indentation_width rule has to be enabled manually since its an opt_in rule. Also the format for the configuration was no longer correct with the current version of swiftlint.

Also the rule isn't perfect. It flags a few issues which are not necessarily incorrect indentation, which can be seen in the diff for issues that were fixed. Still I think its better to have it on than off since we had quite a few indentation issues.

In this regard the SwiftFormat tool is better at detecting incorrect indentation since it didn't flag these correctly indented lines as issues which I had to fix for SwiftLint.

Note: We don't necessarily have to fix these one off indentation issues if we get a warning about them. For example in a multi-clause guard statement, having the guard on a line by itself may feel unusual but I actually like the way this looks. Alternatively we can add a // swiftlint:disable:this indentation_width to the few weird issues that due pop up.

Reviewed By: joesus

Differential Revision: D30003594

fbshipit-source-id: 0ec0aa481cbcef8b081e3aaf08c9884f56939e09