Remove Internal CoreKit Header Dependency - "FBSDKCheckmarkIcon.h"
Summary:
We currently leak internal types and members via an ‘internal’ header that’s used outside of the module. This needs to be fixed so that we have a clean interface to work from.
This means removing imports from FBSDKCoreKit+Internal.h until it can be removed entirely.
Additional reasons to make this change:
Having an internal umbrella header is contrary to the goal of more modular code because it creates artificial coupling.
The header is used outside of CoreKit. This is just not okay anymore. We need to have clear modular boundaries as a prerequisite to rearchitecting or rewriting the SDK.

Reviewed By: joesus

Differential Revision: D30256470

fbshipit-source-id: 2d970395be2a2950f47ba8e2cee9f6122bf24a28