Fix nested swiftlint files not being picked up
Summary:
According to swiftlint's documentation

> If you want to use use nested configurations, you can't use the --config parameter

We use the --config parameter since our .swiftlint.yml file is in the base ios-sdk directory which is one level up from where swiftlint is run from the various Xcode projects.

This fixes the issue by using "cd .." to change the working directory to be one level up and then running "swiftlint {dir_name}" so that swiftlint is only run on files in that directory.

(Longer term we should update so that it only runs on changed files. This will have to work with both hg and git.)

Differential Revision: D30396094

fbshipit-source-id: 6eb37fa38134d103dfdcbf6bd6034158c45911fe