Depcrating Game Request Payload Delegate Method
Summary:
The functionality for a game request payload stays the same except for a new delegate method with a name that makes more sense or distinguish it from the context token ID method.
In this diff I deprecate delegate method: `updatedURLContaining` and 'gamingRequestID' within GamingPayload
Add a better named the better named method: `parsedGameRequestURLContaining:gameRequestID:`

This seperates the `gamingRequestID` from the GamingPayload and instead returns the `gamingRequestID` in the delegate method.

Differential Revision: D30353935

fbshipit-source-id: 441a5882ea7e6a60aa17186799dd9cdcd0a5210d