Disable the no_grouping_extension rule
Summary:
It's often useful to use grouping extensions for protocol conformance or to group non-test helper methods is XCTest files. For this reason we want to allow grouping extensions.

Note: The other changes are due to autoformatting with the prettier extension. I had been reverting those before but I got tired of doing this and so am just going to commit them.

Differential Revision: D30423651

fbshipit-source-id: 4890bbac7ee830169253030c5c5bffbb71126c49