Fix differences in FBAEMKit's xcconfig settings
Summary:
In other folders such as FBSDKLoginKit/Configurations the Shared folder is a symlink to ../../Configurations.

However for FBAEMKit it looks like the folder was copied over which causes a few discrepencies. I attempted to fix this by symlinking it like other kits in D30205446 but there is an tvOS related failure. I'm not yet sure why the additional tvOS configs that FBAEMKit doesn't use is causing this.

Nevertheless I'm going to fix the discrepencies first and then worry about fixing the symlink after I figure out what about the tvOS configs is causing this.

Reviewed By: wx0165927473

Differential Revision: D30425614

fbshipit-source-id: 9c624bd412ecf838bfb0143cb141f37a9b2da25f