Fix failure to build universal tv kits.
Summary:
This is a little strange. I was seeing a failure locally to build the `BuildAllKits_TV` target. This uses the script `build-universal-tvos-framework.sh`. The script had extra arguments to clean and build and this is what was failing.

The targets are used from CI so there will be as fresh checkout so the clean is redundant anyway. Reasoning that it should be fine to remove the offending lines.

This diff also fixes an unrelated Swiftlint warning.

Reviewed By: jawwad

Differential Revision: D30310232

fbshipit-source-id: 3f970c06804fa0419a4b4346b568813cb948dcd1