Fix duplicate folder reference in FacebookGamingServices
Summary:
When building from the command line this error would appear:

The file reference for "../Tests/FacebookGamingServices/Internal" is a member of multiple groups ("FacebookGamingServices" and "Tests"); this indicates a malformed project.  Only the membership in one of the groups will be preserved (but membership in targets will be unaffected).  If you want a reference to the same file in more than one group, please add another reference to the same path.

Reviewed By: joesus

Differential Revision: D30486205

fbshipit-source-id: 8aeeb2387dbbf0635e7f1bcf230fd6a07e146d79