Fix the SwiftLint Codemod job not using our main .swiftlint.yml
Summary:
SwiftLint can use a nested config but they are not recursive.

It uses the config in the directory from where it is run, and then is uses the closest config to the file that it's running on.

Added some overridden settings that should always be applied to the nested config.

Reviewed By: joesus

Differential Revision: D30461064

fbshipit-source-id: cf7fdb1d70a1ee50b3b23f9cbb8bfe36a8194b85