Use explicit references to Social framework from share dialog
Summary:
We are _partially_ attempting to dynamically reference the Social framework.  It's only partial because we actually have the framework import in the file in addition to other symbols.

This removes the class ~~and service type~~ lookup~~s~~ in favor of just referencing the framework.  The dynamic ShareKit target now explicitly declares its dependency on Social, but _weakly_.

The dynamic framework loading functions are still in place along with their tests, but we should determine whether they are of any use and then address them separately.

Reviewed By: joesus

Differential Revision: D30459032

fbshipit-source-id: 7ec028831644712118295f0620c13d25e9ae22e1