Remove symlinked headers from CoreKit
Summary:
Deletes the symlinks and does the rest of the configuration work to use the moved headers.

The changes to the project files are a little hard to read. High level chart:

CoreKit -> needs to add every file. All header files in the include directory should be public. Any other header files should be project.

CoreKitTV -> needs to add every file. Some of the header files in the include directory should be public, others should be omitted. Some but not all of the other header files in the directory should be project, others should be excluded.

A good next step for this would be to figure out how to best structure it to ignore non-tv files and then we can simplify the configuration quite a bit.

Note: Actually moving the headers in the next commit since otherwise HG will see the changes as additions instead of moved files.

Reviewed By: jawwad

Differential Revision: D30522335

fbshipit-source-id: 2036bc47ab2a8e98f9fbf7173b020d5d5e270694