Adds context size to Gaming Context
Summary:
Adds the missing values of context size to the gaming context.
After a conversation with Nacho and David we decided to move forward with context size and deprecate context type. The reason is in the future all context will have a type of generic. Therefore that value will not be useful for developers.

Differential Revision: D30385314

fbshipit-source-id: eca488c550f80ef56bcc7101ce479bcb91569097