Fix violations of the file_name rule
Summary:
Note: I didn't actually enable it because there are still about 12 issues with file names that don't match their containing classes that aren't as easy to fix. Some of the remaining issues are with mismatched cases and others are extensions.

This is the text of the warning:
File Name Violation: File name should match a type or extension declared in the file (if any). (file_name)

Reviewed By: KylinChang

Differential Revision: D30505132

fbshipit-source-id: 2a3fa95af6734b3a4c3ac987c591fcd41fe11bef