Remove unused imports identified by AppCode
Summary: Used AppCode to identify some unused imports to remove. There were many more that it thought were used but the project would not build without. Reverted any files that were causing an issue completely (for the sake of expediency) instead of checking to see which imports were causing the issue.

Reviewed By: KylinChang

Differential Revision: D30658642

fbshipit-source-id: ea0befcc8bb4bf788d1aa8b40e53f31c41eb8449