Update (almost) all untyped NSDictionary objects to be of type NSDictionary<NSString *, id>
Summary:
There was only one instance that was NSDictionary<NSNumber *, id>

Performed the following steps:
1. Replaced all instances of "NSDictionary " with "NSDictionary<NSString *, id> "
2. Changed NSString to NSNumber in the protocolMap method in FBSDKBridgeAPIRequest to fix compile error.
3. Updated "[AnyHashable: Any]" to "[String: Any]" in Swift to fix compile errors

This reduces the number of warnings for the new PATTERNLINT objc-headers-collection-generics warning from 139 to 50.

While this is not foolproof, in almost all cases a dictionary's key will be a String. If we find an instance in which it is not we can fix it later. Even if some of these are not actually typed with String it won't cause a crash since this is just a type hint.

Also it's nicer to see [String: Any] than [AnyHashable: Any] since we are almost always going to use Strings as keys in dictionaries.

Differential Revision: D30689792

fbshipit-source-id: efd25f74f2a1328718398272359c72fac35b38b3