Attempt to fix unrecognized selector crash in issue #1821
Summary:
Sometimes methods defined on a category won't be included without using the -ObjC linker flag. This class defines category methods on NSError but they don't really need to be defined this way. Changed the class to be FBSDKLoginErrorUtil instead of a category on NSError. FBSDKLoginError is already defined as an enum.

https://developers.facebook.com/docs/ios/troubleshooting#unrecognizedselector

Reviewed By: joesus

Differential Revision: D30684031

fbshipit-source-id: 18b00352245bada47e3ba57910bc944b1ec186f4