Support dependency injection for bridge API request type
Summary:
Add dependency injection to FBSDKBridgeAPIRequest for its class dependencies on:
* internal URL opener
* internal utility
* settings

Note: This was going to be a mechanism for removing OCMock from the share dialog tests, but I'm opting instead for a factory interface for creating these requests.  This was already done and is needed, so I'm submitting it anyway.

Reviewed By: jawwad

Differential Revision: D30675909

fbshipit-source-id: 272e03d62a47f88304fc43f5c71289a0b6721d56