Create a factory interface for bridge API requests
Summary: Bridge API requests should be created through a factory interface.  This diff creates that interface and a concrete implementation that forwards calls to the request type that is currently used for creating requests.  As Core types making requests adopt dependency injection, they can start with using the concrete implementation for production and we can move the implementation that is currently in the request type into the factory once that transition is complete.

Reviewed By: jawwad

Differential Revision: D30682348

fbshipit-source-id: 5101b376a79530f6afbc73c9c93adbe16dd70dd4