Combine LegacyGamingServices module with FacebookGamingServices module
Summary:
Was going to split this up but in the event that we need to roll it back it's easier as a mondo-diff. Also there really isn't a great way to split this. :'(

What to know:

We split GamingServices into three modules so that we could incrementally rewrite in Swift. SPM doesn't allow mixed source targets so this was necessary.

We then realized while working on XCFramework-SPM support that it would be a terrible experience to have to add an extra dependency that wouldn't present until runtime.

ie. As a user wanting to add FacebookGamingServices module from SPM, I do not want to have to include LegacyGamingServices to avoid a runtime crash.

Better to untangle this structure now, build two XCFrameworks (FacebookGamingServices and the wrapper layer FBSDKGamingServicesKit) and have fewer dependencies to provide.

Reviewed By: jawwad

Differential Revision: D30794657

fbshipit-source-id: 9d018cd070a29615e9b4d0478b5e380e0b176c09