Block Passing in Context ID
Summary:
Before we were allowing developers to pass in any context token id and perform a custom update graph request. With this change we will no longer allow that and instead check the current gaming context for the context token ID.

This check occurs when we're converting the local CustomUpdateContent object to the remote object sent to the graph api. This process is apart of the graph request call and will throw if the user is not in a gaming context.

BE: Removed the extra initializer with the required parameters in the CustomUpdateContent object. Instead added default values for the none required parameters in the first initializer.

SIde note: Since we are transitioning to swift the initializer for the Gaming Context is public. Ideally once the transition is over we would make it internal. joesus

Reviewed By: joesus

Differential Revision: D30934237

fbshipit-source-id: 4c78e5844e541f357a878c83aa8033583f429823