Get Tournaments added checks and fix
Summary:
After adding the get tournaments api to hackbook there were some missing checks and a bug.
We need to check to make sure that the user is has an Access token or otherwise the graph API will fail with invalid token. We need to check that they're logged on with Gaming Domain.

There was a bug where the API was actually making a call to "/me" which it should actually be "{userID}/Tournaments".

Reviewed By: samodom

Differential Revision: D31022008

fbshipit-source-id: de726de9c29266d1a60ca829a132cd7661efb295