Abstracting dependencies from AppLinkUtility
Summary:
Abstracts dependencies on:
Settings
AppEventsUtility
AppEventsConfigurationManager

Added a couple of protocols to describe some behaviors of AppEventsUtility:
`AppEventDropDetermining`
`AppEventParametersExtracting`

Note that this uses AppEventsConfigurationManager as an instance. This required a little wiggling but now it should be possible to use it as an instance in other places where it's currently using class methods.

Reviewed By: samodom

Differential Revision: D31064422

fbshipit-source-id: d3feea729a8b23fb47759976787e06e9962eb954