Get Tournaments Decoding Bug
Summary:
In this diff i fix the decoding bug that was in previous diffs. The issue was that the json data and expiration date received from the graph API. We need to parse a json object with a data field with a value of an array of tournament json objects.

The value for an expiration date inside the tournament json is actually an iso 8601 date and not a time interval. So In this diff i created a date formatter that can be used throughout the SDK to format said dates in the future.

Reviewed By: joesus

Differential Revision: D31060671

fbshipit-source-id: b15521cb6a7702122e1d6dc22c2e3e80679ff92c