Remove redundant IPHONEOS_DEPLOYMENT_TARGET settings in various xcconfig files
Summary:
Currently the IPHONEOS_DEPLOYMENT_TARGET is defined in a few different places.
- In the xcodegen_project_common.yml file that xcodegen uses
- In many different xcconfig files

If the deploymentTarget is updated in xcodegen_project_common.yml it updates the IPHONEOS_DEPLOYMENT_TARGET in the pbxproj files however the settings are still overridden by xcconfig files.

https://pxl.cl/1Qs4B

For example for the FBSDKCoreKit-Dynamic target the resolved version is still 9.0 due to the FBSDKCoreKit-Dynamic.xcconfig file which is actually inheriting this setting from the iOS.xcconfig file.

This commit removes the extra redundant settings in the individual xcconfig files.

For now I'm leaving the setting in the following files.
- Configurations/FacebookSDK-Project.xcconfig
- Configurations/Platform/iOS.xcconfig
- xcodegen_project_common.yml

Eventually the setting should be defined in either xcodegen_project_common.yml or an xcconfig file.

Reviewed By: joesus

Differential Revision: D31348431

fbshipit-source-id: f36497e7ec4b811a94eab3c509bc40c19ca90cd0