Fix bad default graph recovery flag in settings
Summary: When adding an instance property for enabling graph recovery, we replaced a "disable" flag having a false default with an "enable" flag having a false default.  This caused the wrong value to be the default with the new property.  This fix sets the new property to the correct true default.

Reviewed By: joesus

Differential Revision: D31545724

fbshipit-source-id: 41b6f2651a5d613f808273a6dcadf28804301e65