Fix project not building in Xcode 13
Summary: Fixed the project not building in Xcode 13 by removing extra code in build-universal-framework.sh that was building a fat binary. For now the script is still needed in order to copy the *.framework files from derived data into the build directory in the project since there are some CI jobs that depend on it.

Reviewed By: joesus

Differential Revision: D31568435

fbshipit-source-id: 9d32cf7d6d53354eaaa62452a05a1d9bdc7df844