Fix Xcode 13 warning: "Building targets in manual order is deprecated"
Summary:
The full warning is:
```
Building targets in manual order is deprecated - choose Dependency Order in scheme settings instead, or set DISABLE_MANUAL_TARGET_ORDER_BUILD_WARNING in any of the targets in the current scheme to suppress this warning
```

It looks like parallelizeBuildables should be set to YES in the scheme for this.

This is controlled by the `parallelizeBuild` option in xcodegen which is true by default. Removed where it was being set to false.

Reviewed By: samodom

Differential Revision: D31194925

fbshipit-source-id: 463a315ad5e9d6121723f5802067e5b293138a26